requirements.txt
astroid==2.3.3
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fonttools==4.39.3
graphviz==0.20.1
grpcio==1.16.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-resources==5.12.0
incremental==16.10.1
ipaddr==2.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.15.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
logilab-common==1.4.3
matplotlib==3.7.1
mccabe==0.6.1
mininet==2.3.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.2
oauthlib==3.1.0
ovs==2.13.3
packaging==23.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.5.0
ply==3.11
prompt-toolkit==2.0.10
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pydot==1.4.2
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyroute2==0.5.9
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scapy==2.4.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.1.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
thrift==0.13.0
traitlets==4.3.3
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.15.0
zope.interface==4.7.1
